update for prob

+++ vision/tp7/src/utils.cpp
@@ -267,11 +267,9 @@ void normalisePPM(float* imageFloat, int h, int w,int a, int b){
 		    for(k=0;k<3;k++){
           imageFloat[indppm(i,j,k,w)]=(((double)imageFloat[indppm(i,j,k,w)]-(double)min)/
                                      ((double)max-(double)min))*(b-a)+(double)a;
-
 		    }
 		}
 	}
-
 }
 
 void normalisePGM(float* imageFloat, int h, int w,int a, int b){
@@ -295,13 +293,12 @@ void normalisePGM(float* imageFloat, int h, int w,int a, int b){
 			              ((double)max-(double)min))*(b-a)+(double)a;
 		}
 	}
-
 }
 
 void floatP2intP(float* tab_f,int* tab_i,int n){
 	for(int i=0;i<n;i++)tab_i[i]=tab_f[i]+0.5;
 }
 
-void intP2floatP(int* tab_i,int* tab_f,int n){
+void intP2floatP(int* tab_i,float* tab_f,int n){
 	for(int i=0;i<n;i++)tab_f[i]=tab_i[i];
 }
